requirements.txt
httpx==0.28.0
fastapi==0.115.13
injector==0.21.0
loguru==0.7.2
motor==3.6.0
pydantic==2.5.3
pydantic_core==2.14.6
python-decouple==3.8
python-dotenv==1.0.0
uvicorn==0.34.3
email-validator==2.0.0
